+++ src/VSPackage.cs
@@ -1,13 +1,15 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
 
 namespace FileDiffer
 {
     [PackageRegistration(UseManagedResourcesOnly = true)]
     [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
-    [Guid(PackageGuids.guidDiffFilesCommandPackageString)]
+    [ProvideAutoLoad(UIContextGuids80.SolutionExists)]
+    [Guid(PackageGuids.guidPackageString)]
     public sealed class VSPackage : Package
     {
         protected override void Initialize()